Add tests for the #body and #to_xml methods
I had previously only been using the #to_xml and #parts methods, so I hadn't
run this this before. The issue was that Nori does not know how to handle an
instance of Mail::Body rather than a string. It seems that the correct thing
to do is to just return a string.

Also brought in some same defaults from Savon::GlobalOptions for this test.
Before it wouldn't have been possible to actually call the #body method in
the tests because we didn't specify how to take the response body tags and
turn them into snakecase symbols